# docs/swagger.yaml
basePath: /
definitions:
  github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: Success
        type: string
      meta:
        $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.PaginationMeta'
      status:
        example: success
        type: string
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.BujpResponse:
    properties:
      address:
        example: Jl. Kebun Mas Blok Q147
        type: string
      bjb_account_number:
        example: "0123456789"
        type: string
      code:
        example: BUJP001
        type: string
      created_at:
        type: string
      email:
        example: info@bujp.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: GARDA MUTIARA TIMUR
        type: string
      phone:
        example: "081234567890"
        type: string
      pic_name:
        example: John Doe
        type: string
      pic_phone:
        example: "081234567890"
        type: string
      pks_document:
        example: bujps/pks/doc.pdf
        type: string
      pks_document_url:
        example: https://s3.example.com/presigned-url
        type: string
      pks_number:
        example: PKS/BJB/2026/001
        type: string
      pks_validity_end:
        example: "2027-01-01"
        type: string
      pks_validity_start:
        example: "2026-01-01"
        type: string
      status:
        example: active
        type: string
      updated_at:
        type: string
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.CreateMonthlyReportRequest:
    properties:
      bujp_id:
        type: string
      period:
        description: YYYY-MM
        type: string
      report_file:
        type: string
      total_absent:
        minimum: 0
        type: integer
      total_guards:
        minimum: 0
        type: integer
      total_locations:
        minimum: 0
        type: integer
      total_patrols:
        minimum: 0
        type: integer
      total_payroll:
        minimum: 0
        type: number
      total_permission:
        minimum: 0
        type: integer
      total_present:
        minimum: 0
        type: integer
      total_sick:
        minimum: 0
        type: integer
    required:
    - bujp_id
    - period
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.GenerateMonthlyReportRequest:
    properties:
      bujp_id:
        type: string
      force:
        type: boolean
      period:
        description: YYYY-MM
        type: string
    required:
    - period
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.MonthlyReportResponse:
    properties:
      bujp:
        $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.BujpResponse'
      bujp_id:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      creator:
        $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.UserResponse'
      id:
        type: string
      period:
        type: string
      report_file:
        type: string
      total_absent:
        type: integer
      total_guards:
        type: integer
      total_locations:
        type: integer
      total_patrols:
        type: integer
      total_payroll:
        type: number
      total_permission:
        type: integer
      total_present:
        type: integer
      total_sick:
        type: integer
      updated_at:
        type: string
    type: object
  github_com_ganiramadhan_ganipedia_backend_internal_model.PaginationMeta:
    properties:
      has_next:
        example: true
        type: boolean
      has_previous:
        example: false
        type: boolean
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.UpdateMonthlyReportRequest:
    properties:
      period:
        type: string
      report_file:
        type: string
      total_absent:
        minimum: 0
        type: integer
      total_guards:
        minimum: 0
        type: integer
      total_locations:
        minimum: 0
        type: integer
      total_patrols:
        minimum: 0
        type: integer
      total_payroll:
        minimum: 0
        type: number
      total_permission:
        minimum: 0
        type: integer
      total_present:
        minimum: 0
        type: integer
      total_sick:
        minimum: 0
        type: integer
    type: object
...
  github_com_ganiramadhan_ganipedia_backend_internal_model.UserResponse:
    properties:
      bujp:
        $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.BujpResponse'
      bujp_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      created_at:
        type: string
      email:
        example: john@example.com
        type: string
      full_name:
        example: John Doe
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      last_login:
        type: string
      phone:
        example: "081234567890"
        type: string
      photo:
        example: users/photo.jpg
        type: string
      role:
        example: guard
        type: string
      status:
        example: active
        type: string
      updated_at:
        type: string
    type: object
host: localhost:4000
info:
  contact:
    email: support@hrmis.com
    name: API Support
  description: HRMIS Backend API - Human Resource Management Information System
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: HRMIS API
  version: "1.0"
paths:
...
  /api/v1/monthly-reports:
    get:
      description: Get list of monthly reports with pagination and filters
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by BUJP ID
        in: query
        name: bujp_id
        type: string
      - description: Filter by period (YYYY-MM)
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.MonthlyReportResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all monthly reports
      tags:
      - monthly-reports
    post:
      consumes:
      - application/json
      description: Create a new monthly report
      parameters:
      - description: Monthly report data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.CreateMonthlyReportRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.MonthlyReportResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Create new monthly report
      tags:
      - monthly-reports
  /api/v1/monthly-reports/{id}:
    delete:
      description: Delete monthly report by ID
      parameters:
      - description: Monthly Report ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete monthly report
      tags:
      - monthly-reports
    get:
      description: Get monthly report details by ID
      parameters:
      - description: Monthly Report ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.MonthlyReportResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Get monthly report by ID
      tags:
      - monthly-reports
    put:
      consumes:
      - application/json
      description: Update monthly report by ID
      parameters:
      - description: Monthly Report ID
        in: path
        name: id
        required: true
        type: string
      - description: Monthly report data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.UpdateMonthlyReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.MonthlyReportResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Update monthly report
      tags:
      - monthly-reports
  /api/v1/monthly-reports/generate:
    post:
      consumes:
      - application/json
      description: Calculate aggregated stats for a (BUJP, period) and persist a monthly
        report. Non-superadmin callers are clamped to their own BUJP.
      parameters:
      - description: Generate request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.GenerateMonthlyReportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.MonthlyReportResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_ganiramadhan_ganipedia_backend_internal_model.APIResponse'
      security:
      - BearerAuth: []
      summary: Generate monthly report
      tags:
      - monthly-reports
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"